Move terminal completion item impl into its own file
